api/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
requests==2.32.3
python-dotenv==1.0.1
numpy==1.26.4
pydantic==2.10.0
mangum==0.17.0
